refactor: move AgentWithRelations type import to a dedicated types file
Updated imports for AgentWithRelations across multiple components and services to reference the new types file, improving code organization and maintainability.
