+++ web-app/src/app/(landing)/agents/page.tsx
@@ -2,8 +2,9 @@ import { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 
 import { Agents, AgentsNotAvailable } from "@/components/agents";
-import { AgentWithRelations, getAgents } from "@/lib/db/services/agent.service";
+import { getAgents } from "@/lib/db/services/agent.service";
 import { calculateAgentHumandReadableCreditCost } from "@/lib/db/services/credit.service";
+import { AgentWithRelations } from "@/lib/db/types/agent.types";
 
 import { FeaturedAgent } from "./components/featured-agent";
