+++ web-app/src/app/app/agents/components/filtered-agents.tsx
@@ -9,8 +9,8 @@ import {
   AgentsNotFound,
 } from "@/components/agents";
 import { getTags } from "@/lib/db/extension/agent";
-import { AgentWithRelations } from "@/lib/db/services/agent.service";
 import { AgentListWithAgent } from "@/lib/db/services/agentList.service";
+import { AgentWithRelations } from "@/lib/db/types/agent.types";
 
 import { FilterState } from "./use-gallery-filter";
