+++ web-app/src/components/breadcrumb-navigation/breadcrumb-navigation.client.tsx
@@ -13,7 +13,7 @@ import {
   BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb";
 import { getName } from "@/lib/db/extension/agent";
-import { AgentWithRelations } from "@/lib/db/services/agent.service";
+import { AgentWithRelations } from "@/lib/db/types/agent.types";
 
 interface BreadcrumbSegment {
   label: string;
